requirements.txt
ccxt>=4.2.75
python-dotenv>=1.0.0
